feat: move Tag to a separate package
