+++ package.json
@@ -68,8 +68,10 @@
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
     "@contentful/browserslist-config": "^1.0.1",
-    "@parcel/babel-plugin-transform-runtime": "2.0.0-nightly.1823",
-    "@parcel/babel-preset-env": "2.0.0-nightly.526",
+    "@parcel/babel-plugin-transform-runtime": "2.0.0-nightly.2233",
+    "@parcel/babel-preset-env": "2.0.0-nightly.612",
+    "@parcel/transformer-typescript-types": "2.0.0-nightly.612",
+    "@parcel/packager-ts": "2.0.0-nightly.612",
     "@storybook/addon-a11y": "^6.1.11",
     "@storybook/addon-essentials": "^6.1.11",
     "@storybook/addon-storysource": "^6.1.11",
@@ -124,7 +126,7 @@
     "jest-axe": "4.1.0",
     "jest-watch-typeahead": "0.6.1",
     "lerna": "^3.22.1",
-    "parcel": "2.0.0-nightly.524",
+    "parcel": "2.0.0-nightly.610",
     "plop": "^2.5.3",
     "postcss": "^8.2.2",
     "postcss-copy-assets": "^0.3.1",
