+++ packages/components/tag/stories/Tag.stories.tsx
@@ -1,17 +1,73 @@
 import React from 'react';
-import type { Meta, Story } from '@storybook/react/types-6-0';
 
-import { Tag } from '../src/Tag';
-import type { TagProps } from '../src/Tag';
+import { Tag, TagProps } from '../src/Tag';
+import SectionHeading from '../../../forma-36-react-components/src/components/Typography/SectionHeading/SectionHeading';
+import Flex from '../../../forma-36-react-components/src/components/Flex/Flex';
 
 export default {
+  title: 'Components/Tag',
   component: Tag,
   parameters: {
-    propTypes: Tag['__docgenInfo'],
+    propTypes: [Tag['__docgenInfo']],
   },
-  title: 'Components/Tag',
-} as Meta;
-
-export const Default: Story<TagProps> = (args) => {
-  return <Tag {...args}>Tag</Tag>;
+  argTypes: {
+    className: { control: { disable: true } },
+    testId: { control: { disable: true } },
+  },
 };
+
+interface Args extends TagProps {
+  tagText?: string;
+}
+
+export const basic = ({ tagText, ...args }: Args) => (
+  <Tag {...args}>{tagText}</Tag>
+);
+
+basic.args = {
+  tagType: 'primary',
+  tagText: 'Published',
+};
+
+export const overview = () => (
+  <>
+    <Flex marginBottom="spacingS">
+      <SectionHeading element="h3">Tag types overview</SectionHeading>
+    </Flex>
+    <Flex marginBottom="spacingM" alignItems="center">
+      <Flex marginRight="spacingS">
+        <Tag tagType="primary">primary</Tag>
+      </Flex>
+    </Flex>
+    <Flex marginBottom="spacingM" alignItems="center">
+      <Flex marginRight="spacingS">
+        <Tag tagType="primary-filled">primary filled</Tag>
+      </Flex>
+    </Flex>
+    <Flex marginBottom="spacingM" alignItems="center">
+      <Flex marginRight="spacingS">
+        <Tag tagType="positive">positive</Tag>
+      </Flex>
+    </Flex>
+    <Flex marginBottom="spacingM" alignItems="center">
+      <Flex marginRight="spacingS">
+        <Tag tagType="negative">negative</Tag>
+      </Flex>
+    </Flex>
+    <Flex marginBottom="spacingM" alignItems="center">
+      <Flex marginRight="spacingS">
+        <Tag tagType="warning">warning</Tag>
+      </Flex>
+    </Flex>
+    <Flex marginBottom="spacingM" alignItems="center">
+      <Flex marginRight="spacingS">
+        <Tag tagType="muted">muted</Tag>
+      </Flex>
+    </Flex>
+    <Flex marginBottom="spacingM" alignItems="center">
+      <Flex marginRight="spacingS">
+        <Tag tagType="secondary">secondary</Tag>
+      </Flex>
+    </Flex>
+  </>
+);
