+++ packages/forma-36-react-components/package.json
@@ -33,6 +33,7 @@
     "@babel/runtime": "^7.3.4",
     "@contentful/forma-36-fcss": "^4.0.0",
     "@contentful/f36-tokens": "^4.0.0",
+    "@contentful/f36-tag": "^4.0.0",
     "@popperjs/core": "^2.4.4",
     "classnames": "^2.2.6",
     "csstype": "^3.0.5",
