+++ packages/forma-36-react-components/src/components/Button/Button.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Button, { ButtonProps } from './Button';
 import Flex from '../Flex/Flex';
 import SectionHeading from '../Typography/SectionHeading';
-import Tag from '../Tag';
+import { Tag } from '@contentful/f36-tag';
 import { iconName } from '../Icon/constants';
 
 import notes from './README.mdx';
