--- packages/forma-36-react-components/src/components/Tag/index.ts
@@ -1,2 +1,0 @@
-export * from './Tag';
-export { default } from './Tag';
