+++ packages/forma-36-react-components/src/index.ts
@@ -39,7 +39,8 @@ export { TableRow } from './components/Table/TableRow/TableRow';
 export { ToggleButton } from './components/ToggleButton/ToggleButton';
 export { AssetCard } from './components/Card/AssetCard/AssetCard';
 export { Asset, AssetIcon } from './components/Asset';
-export { Tag } from './components/Tag/Tag';
+export { Tag } from '@contentful/f36-tag';
+export type { TagProps, TagType } from '@contentful/f36-tag';
 export { Heading } from './components/Typography/Heading/Heading';
 export { Modal } from './components/Modal/Modal';
 export { ModalConfirm } from './components/Modal/ModalConfirm/ModalConfirm';
