+++ scripts/plop-templates/package/package.json.hbs
@@ -19,7 +19,7 @@
     "url": "https://github.com/contentful/forma-36"
   },
   "scripts": {
-    "build": "parcel build index.ts --no-minify"
+    "build": "parcel build index.ts"
   },
   "dependencies": {
     "@babel/runtime": "^7.6.2",
